Fixed timestamp size

+++ erizo/src/erizo/media/ExternalOutput.h
@@ -50,7 +50,7 @@ namespace erizo{
       int prevEstimatedFps_;
       int warmupfpsCount_;
       int sequenceNumberFIR_;
-      unsigned long lastTime_;
+      unsigned long long lastTime_;
 
       int video_stream_index, bufflen, aviores_, writeheadres_;
 
@@ -67,7 +67,7 @@ namespace erizo{
       unsigned char deliverMediaBuffer_[3000];
       unsigned char unpackagedBuffer_[15000];
       unsigned char unpackagedAudioBuffer_[15000];
-      unsigned long initTime_;
+      unsigned long long initTime_;
   };
 }
 #endif /* EXTERNALOUTPUT_H_ */
